closes issue #624

+++ src/main/java/servlets/module/challenge/DirectObjectBankLogin.java
@@ -3,6 +3,7 @@ package servlets.module.challenge;
 import dbProcs.Database;
 import java.io.IOException;
 import java.io.PrintWriter;
+import java.math.BigInteger;
 import java.sql.CallableStatement;
 import java.sql.Connection;
 import java.sql.ResultSet;
@@ -141,7 +142,7 @@ public class DirectObjectBankLogin extends HttpServlet {
       ResourceBundle errors)
       throws SQLException {
 
-    float currentBalance = getAccountBalance(accountNumber, applicationRoot);
+    long currentBalance = getAccountBalance(accountNumber, applicationRoot);
     String bankForm =
         "<h2 class='title'>"
             + bundle.getString("bankForm.yourAccount")
@@ -233,7 +234,7 @@ public class DirectObjectBankLogin extends HttpServlet {
     ResourceBundle bundle =
         ResourceBundle.getBundle("i18n.servlets.challenges.directObject.directObjectBank", locale);
 
-    float currentBalance = getAccountBalance(accountNumber, applicationRoot);
+    long currentBalance = getAccountBalance(accountNumber, applicationRoot);
     String bankForm =
         "<h2 class='title'>"
             + bundle.getString("bankForm.yourAccount")
@@ -317,18 +318,18 @@ public class DirectObjectBankLogin extends HttpServlet {
    * @return Returns a Float Value representing the balance
    * @throws SQLException If no rows found or if SQL error occurs
    */
-  public static float getAccountBalance(String accountNumber, String applicationRoot)
+  public static long getAccountBalance(String accountNumber, String applicationRoot)
       throws SQLException {
     Connection conn = Database.getChallengeConnection(applicationRoot, "directObjectBank");
     CallableStatement callstmt;
-    float toReturn = 0;
+    long toReturn = 0;
     try {
 
       callstmt = conn.prepareCall("CALL currentFunds(?)");
       callstmt.setString(1, accountNumber);
       ResultSet rs = callstmt.executeQuery();
       if (rs.next()) {
-        toReturn = rs.getFloat(1);
+        toReturn = rs.getLong(1);
       } else {
         throw new SQLException("Could not Get Funds. No Rows Found From Query");
       }
